src/list_projects/requirements.txt
google-cloud-storage
gcloud-flask-oauth-cors